Return a more accurate error when non-existing repository records
DeleteRepository deletes a repository's database records. It deletes
the repository's entry in the `repositories` table which marks the
repository as having been deleted. It also deletes the primary's record
in the `storage_repositories` table to indicate the primary's replica
has been deleted. Right now it returns a repository doesn't exist error
if neither of those database records exist. The error isn't really accurate
as it always includes the storage as well which isn't relevant if the
repository doesn't exist at all. With transactional deletes, this is going
to be even more inaccurate as we are deleting records for multiple storages
in one go. To make this a bit nicer, this commit changes the returned error
to be a more generic no rows affected error.