operations: Use object quarantine directories in UserCreateTags
When the UserCreateTag RPC fails, it may leave behind the newly created
tag object in the repository's object database, even though it's not
referenced by anything. This is typically not much of a problem given
that it'll get cleaned up eventually, but this is a general problem our
RPCs have which manually create objects: even though new objects may not
conform to what we do want to accept into a repository, it will end up
in the object directory anyway.

Fix this shortcoming by using our new infrastructure to create manual
object quarantine directories. This quarantine directory will host all
newly created objects. If the RPC fails, it will simply get discarded,
otherwise its objects will be migrated into the target repository.

Furthermore, this has the benefit that we can now pass information about
the quarantine to Rails. Like this, it can directly enumerate all new
objects via the quarantine directory instead of having to do a graph
walk, which can be a huge speedup depending on the repository shape and
size.

Using UserCreateTag as the first RPC to use quarantine directories can
be attributed to the fact that it's one of the few RPCs which use Git to
create objects instead of libgit2. Other RPCs will be converted at a
later point if use of the quarantine directory delivers the expected
results.

Changelog: changed