transactions: Simplify computation of voter state
In order to determine the voter state in a transaction, we iterate
through all subtransactions and soak up the result in there. But given
that subtransactions are always created with all voters of the
transaction, we know that on each iteration, we'll override all results
of the preceding subtransaction anyway. Ultimately, we thus end up with
the state as recorded by the last subtransaction, and it's clear that
the iteration is pointless.

Simplify the code to just take the state of the last subtransaction.