Relocate GitLab HTTP test server from testhelper package
The testhelper package contains too many unrelated stuff that
needs to be moved elsewhere closer to the actual usage or
real implementation. In some cases it can help to avoid
circular dependencies between packages in the tests.
As well it will give us more consistent and structured codebase.

The types and functions moved into gitlab package got a
rename by removing Gitlab prefix because it causes a lint issue
on double naming (because of the same package name).

A couple of structs such as postReceiveResponse that are copies
of the unexported gitlab package structs were removed as they
are not used anymore.