Fix error check to get more info
There is no sense in use of the stderr and stdout if Run
on Command fails with an error due to usage of the require.NoError.
require.NoError interrupts execution of the test immediately, so
we won't get values of the stdout and stderr to inspect.
That is why we change the order of the checks and replaces require
to assert for the stream verification.