tempdir: Split implementation into tempdir and storage specific items
The tempdir implementation is currently hosted in a single file, but
it's in fact got two different components: first the tempdir-related
code itself which is concerned about a single tempdir and its lifecycle.
And second the broader view of how tempdirs are managed at the storage
level.

Split up the implementation into two files to make it more focussed.