catfile: Refactor ParseCommit to accept object ID
The `ParseCommit()` function requires the object ID as input such that
it can fill in the ID field of the resulting `GitCommit` protobuf
message, but it receives as input a full-blown object info. This makes
it less universally useful given that not every calling context does
have object info available.

Refactor the interface to only get an object ID as input.