git: Mark git-for-each-ref(1) as supporting `--end-of-options`
While git-for-each-ref(1) is currently marked as not supporting the
`--end-of-options` separator, it is perfectly capable of supporting it.

One callsite that keeps us from using the separator though is a custom
command implementation for git-for-each-ref(1), which has been
introduced via e54e89905 (Add PostArgFlags, 2019-10-22). This
implementation overrides the validation of arguments such that we can
use flags after positional arguments, which is of course incompatible
with `--end-of-options`. It's unclear why this is done though: the only
callsite of this custom implementation uses it to move the `--contains`
flag after the other reference pattern argument, but git-for-each-ref(1)
is perfectly happy to have that flag as a leading flag.

Remove the custom implementation of the `git.Cmd` interface. With this
code gone, we can now mark the command as supporting `--end-of-options`.