git: Mark git-clone(1) as supporting `--end-of-options`
The git-config(1) command is currently labelled as not supporting the
`--end-of-options` separator, even though Git supports this just fine.

There is currently only a single callsite which keeps us from marking
the command as supporting `--end-of-options`, which is the
implementation of `DeleteConfig()`. What we do there is that we abuse a
`git.ValueFlag` to pass `--unset-all $value`, even though this is in
fact a flag plus a positional argument. This causes Git to misinterpret
the command, where it thinks that `--end-of-options` is a value pattern
to use for `--unset-all`.

Fix the callsite to pass the value as a proper positional argument and
mark the command as supporting the separator.