git: Mark git-ls-tree(1) as supporting `--end-of-options`
The git-ls-tree(1) command is currently labelled as not supporting the
`--end-of-options` separator, even though Git supports this just fine.
What it doesn't support though is intermixed use of this separator with
the `--` separator, but this is exactly how we use it in one of the
callsites.

Adjust the callsites to not pass the revision as a post-separator
argument, but instead as a normal argument. By marking this command as
supporting `--end-of-options`, this has exactly the same semantics given
that both separators can be used interchangeably in this context, but it
now automatically injects the separator.