Makefile: Add help for targets and variables
Our Makefile has quite a lot of different targets and variables which
can be executed and tweaked by the user. Not all of these are self
explanatory, and discoverability of these is abysmal if one doesn't know
the ropes of the Makefile already. To improve the situation, this commit
adds a new target "help", which will print the most important targets
and variables.

One of the issues with such a "help" target is that it would quickly get
out of date if it were to carry documentation itself. Instead, we use a
neat hack to generate documentation from the Makefile itself via awk:
targets or variables preceded by "## "-style comments will have their
comments extracted from the Makefile and will then be printed. Like
this, they are self documenting and easier to keep up to date.