git2go: Pass binary path to runWithGob
We're about to migrate all subcommands of gitaly-git2go to be
implemented on top of the `Executor` structure. Given that this
structure only has paths to binaries instead of the complete config,
this would be hard to do for those which use `runWithGob()`, which
currently requires a complete `config.Cfg` as input.

Convert `runWithGob()` to require the path to the binary, only.
Eventually, this parameter will go away anyway given that we'll make it
a receiver function of the `Executor`, but it helps us to refactor the
code one by one while that's not yet the case.