git2go: Convert executor to receive `config.Cfg` as input
We're about to convert all Git2Go subcommands to use the `Executor`
structure, which will add quite some callsites where we need to create a
new instance of this structure. Constructing it requires some knowledge
on which parameters it needs from the `config.Cfg` though, which is a
bit of a chore.

Refactor its constructor to get a `config.Cfg` instead of the binary
paths, which makes it easier to reason about correctness here.

While at it, rename the constructor from `New()` to `NewExecutor()`: the
package name doesn't make it sufficiently clear that the returned value
will be an executor, so the long name is preferable here.