git2go: Inject repository into subcommands
In order to support use of alternative object directories for a
repository, we'll need to inject it into subcommands such that we can
extract relevant information.

Add a `repository.GitRepo` parameter to all subcommands to enable for
this. No functional change is expected from this commit given that we do
not yet use the repository.