operations: Implement squashing without worktrees
When squashing a range of commits, we create a diff and then apply this
diff on top of the start commit in a separate worktree. This logic is
quite complex and regularly causes failures in case there are binary
files in the diff. Now that we assert that start and end commit of the
squash form a line of heritage, a simple observation is that the
squashed commit's tree is always going to have the exact same contents
as the start commit with the patch applied.

Refactor the code based on this observation to use git-commit-tree(1)
instead of a worktree. The parent of the resulting commit is the start
commit, while the tree of the commit is the end commit's tree.

Note that this also moves the check whether the end commit is an
ancestor of the start commit such that it is guarded by the new feature
flag. This is done such that we can quickly disable this new restriction
in case it turns out that our assumption was wrong here.

Changelog: performance