backup: support of the Cloud storages
After adding a new sinks for the backups in 643d6a084
(StorageServiceSink to support AWS, Azure, Google storage
for the backups, 2021-07-07) we now can integrate them
into the backup task.
The URL for the files to store is the path provided as a
flag to the binary. It should include a specific schema
prefix if a service storage needs to be used (it also
requires additional env vars to be set). The provided path
will be used as a bucket on the storage service where
all backups will be placed.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3633

Changelog: added