Add pagination to GetTreeEntries RPC
If a page limit larger than 0 is set, this RPC will only retrieve that
number of entries from the list.

If a page token is set, this RPC will return entries after this OID.

This reduces the amount of recursive calls needed to flatten a path.

Relates to https://gitlab.com/gitlab-org/gitaly/-/issues/2704

Changelog: added