replication: Await for backlog processing to complete in test
In our tests we start a separate goroute that processes
replication event jobs. Once the test done the cancel
function is called that cancels the context. At this point
the execution of the test is over, but there is a goroutine
still running that processes replication events. It leads
to situations when one test is started, but the previous test
still uses replication queue because of the running background
goroutine. This is not an issue for the in-memory implementation
of the replication queue, but it causes problems for Postgres
based implementation. The fix is await for the goroutine to
complete before the test ends.