ref: Add a testcase with duplicate annotated tags
Annotated tags have two different sources of tag names: first the tag
name as encoded in the tag object itself, and second the name of the
reference that points to the tag object. Given that the former can be
ambiguous, we typically return the latter as name when listing tag
objects. But with the recent conversion of UserCreateTag to use the new
gitpipe package, this detail got lost and we always return the tag name
as encoded in the object.

Add a testcase which demonstrates that we indeed do return ambiguous tag
names for annotated tags.