ref: Fix FindAllTags returning tag names instead of ref names
With the recent port of FindAllTags to the new gitpipe package, we can
now have situations where we return duplicate tag objects. This can
happen when two tag references point to the same annotated tag object:
we return the tag name as encoded in the object in contrast to the tag
name as encoded by the reference. This is a change in behaviour compared
to the old code, which always returned the tag name as encoded by the
reference.

Fix the issue by always using the reference's tag name.

Changelog: fixed