localrepo: Improve type safety of WriteTag/FormatTag functions
Both `WriteTag()` and `FormatTag()` receive the user name and email as
byte slices. This makes it easy to screw up the order of arguments given
that tag name and body are also passed as byte slices.

Improve this by accepting a `gitalypb.User` as argument instead of
separate name and email.