Fix mirroring branch called 'tag' in UpdateRemoteMirror
The Ruby implementation of UpdateRemoteMirror didn't handle references
by their fully qualified names. This caused the mirroring to fail when
attempting to push the tag with `git push <remote> tag` as command
expected the tags name to follow. The Go port mirrored the behavior to
keep the existing tests passing. Now that the feature flag is removed,
we can drop this behavior and allow the mirroring to work regardless of
the branch name.

Changelog: fixed