Fix mirroring branch and tag of same name in UpdateRemoteMirror
UpdateRemoteMirror's Ruby implementation did not use fully qualified names
when pushing references to the mirror. If a branch and tag had the same name,
the resulting git push command would be `git push <remote> <name> <name>`. Git
doesn't allow for this and fails with 'src refspec master matches more than one'.
The Go port of the RPC matched the behavior in order to keep compatiblity with the
tests. Now that the feature flag is removed, we can drop this behavior and allow
pushing to work even if a tag and branch share the same name as the Go implementation
pushes using fully qualified ref names.

Changelog: fixed