Fix mirroring empty repository in UpdateRemoteMirror
UpdateRemoteMirror's Ruby implementation did not handle the case when
there were no branches in the repository that source repository when
mirroring and failed when trying to access its default branch. The RPC
then failed with a NoMethodError when trying to invoke a method on a
Nil object. The Go implementation matched the behavior to pass the existing
tests. Now that the Go implementation is enabled always, we can drop
this behavior and allow the repository mirroring to work even if there
are no branches in the source repository.

Changelog: fixed