storage: Extend locator interface to know about various directories
Knowledge on temporary, cache and state directories is currently part of
the tempdir package. Ideally, determining those paths should be the
responsibility of implementation of the locator interface though, given
that it manages all paths of a given storage. The current design thus
feels like it violates separations of concerns.

Extend the locator interface to prepare for a fix. This will allow us to
convert callers of the tempdir package. While this change does feel
correct by itself, it will also help avoid a dependency cycle at a later
point in this patch series.