tempdir: Convert to use the locator
Convert the tempdir package to use the locator interface to determine
locations of cache and state directories. This will eventually allow us
to get rid of the old interface.

Note that with the new design, we can also easily get rid of a
conceptual weakness in our current test design. Previously, we created a
temporary directory up front in `testMain()` which then got used by the
various test utilities. Now that tempdir by necessity depends on a
locator and thus on an existing storage, this is converted to use ad-hoc
temporary directories as created by our test infrastructure when we set
up new storages.