Fix UpdateRemoteMirror force pushing over diverged refs due to a race
UpdateRemoteMirror currently may force push over diverged refs even if
KeepDivergentRefs is set in the request. This can happen as the RPC
first checks the refs on the remote to see if they have diverged or not.
After these checks, it then force pushes to all refs that have not
diverged. If the refs diverged on the mirror between the RPC checking
them and the force push, the push would overwrite the diverged refs.
This commit prevents accidentally force pushing over the diverged refs
by doing a normal push instead of a force push if KeepDivergentRefs is
set. This then causes the RPC to fail instead of overwriting the refs if
the race manifested.

Changelog: fixed