quarantine: Fix calling RPCs with manual object quarantine environments
When calling the `GetObjectDirectorySize()` RPC with a quarantined
repository, then we currently fail with an error because the quarantine
object directory looks like an attempted path traversal to us. This is
because manual object quarantine directories are not located beneath the
repository, but instead in the storage's temporary directory. The main
intent of this is to get cleanup for free via the tempdir package.

Now that the preceding commit has changed the way we create quarantine
directories to contain a repository-specific prefix, we can easily
determine whether a given quarantine directory's path does indeed belong
to a given repository by simply checking whether the expected quarantine
directory prefix matches the actual path we've got.

Implement this logic for `GetObjectDirectoryPath()`, which fixes calls
to `GetObjectDirectorySize()`.

Changelog: fixed