Remove redundant test setup
The package internal/praefect/service/info has no tests in it
and there is no reason to do a test setup with testhelper.Configure()
call and verification of absent child processes.
The race on the wg from internal/command package can't be reproduced
and actually there is nothing that could lead to such situation:
the WaitGroup.Add method is called before any WaitGroup.Wait method.
The issue raised only once, so maybe it kind of a lag or so.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3703