config: Refactor locator tests
Due to dependency cycles, tests in the config module cannot make use of
any of our testing infrastructure like the testhelper or gittest. As a
result, the test for the locator is hand-crafting a pseudo storage and
then uses it to drive the test. It goes without saying that this pseudo
storage doesn't really match what we'd see in production systems, and it
is hard to reason about the current test setup.

Put the locator tests into a separate "config_test" module. We don't use
any internal symbols anyway, and like this we can refactor the test to
use the usual testing infrastructure.