storage: Move module into `internal/gitaly`
Except for a single use in Praefec's tests, the "storage" module is only
used by Gitaly. Despite that, it's still located at `internal/storage`
and thus indicates that it may be a generic component for reuse.

Move it into `internal/gitaly` to clarify that it is indeed an
implementation detail of Gitaly, only.