operations: Make UserMergeBranch return PreconditionFailed on conflicts
Changing error code from Unknown to PreconditionFailed if there are
merge conflicts that prevent automatic merge on the branch.
It is more correct API behavior as error is known. By changing the
error reason we reduce amount of Unknown errors that are affecting
our SL{A,O}s.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3195

Changelog: changed