remote: Remove hardcoded branch name from error
Currently if an existing branch/tag we're mirroring matches an existing
reference the error message will always refer to 'master'. This is
confusing to users as there's no guarantee this is the problem ref.

This commit updates the error message to use the name of the reference
that triggered the failure.

The error message checked in
'fails_if_tag_and_branch_named_the_same_update_remote_mirror_test' is
shorted to end at 'refs/' because the failed ref returned is
non-deterministic. Either 'refs/heads/master' or 'refs/tags/master' may
be returned here.

Changelog: fixed